Moved the javascript called in the base template to its own javascript file.
Before the javascript code that drives the top bar search was embed in
the template file. Part of the code was also copied in three different
places. This commit removed the duplicated code and fixes some
undefined errors introduced by poor copping.

This code was also cleaned up to follow jslint.