Call the template in the RowExpander with the json
This commit changes to RowExpander to call the row template with the
json property of the record object. Original the template was called
with the data property but we need info in the json property. I was
also unable to figure out how to append the record.data property
itself.