Remove unused data, give the rendering of the abstract to a method
Based on feed back from Chris we need to remove some of the links from
the default row expander.

Also, the abstract field is black when it is a string that equals
'<p></p>'. So in order to filter out abstracts that are "null" we need
to handle the rendering of that filed in the templateLibbary object